Begun Development of Header+Homepage
Yesterday I was at it for 5 hours and achieved literally nothing because no changes I made to any css or php files were affecting the website somehow.. So today, right before I had another go I decided to load a fresh copy of underscores and somehow I was finally able to start changing things. So I've created a header which includes the logo image as well as the links specified in the information architecture document. I changed the layout of the homepage to display a 'static page', and I created a 'homepage' (the page itself is actually untitled, because I'm too tired to try and figure out how to hide the title at the moment) which will obviously be displayed on the homepage. I then added an image gallery to the 'homepage' page, and included three images, these images link to the 'menu' page.