Header and home page adjustments
Made changes to the header:
- removed pages we didnt need (apply now, shop)
- added separate contact us menu item
- made duplicates of all menu items except home and contact us

The reason for the duplicates, is because I took Broderick's content he created, and experimented with putting it all on the one home page.

Home page:
- as mentioned above, took Broderick's content he created and placed it all on the home page, wrapped each page's section up in their associated div tag (i.e. Services section is wrapped in services_content)
- styled each section a tiny amount

As mentione a couple of weeks ago, I was having troubles with CSS caching, and implemented the fix that Greg linked me to which was to disabe CSS caching on every page by adding a random query string. I will remember to remove this before assignment submission.